Optimizing interpreter
So...the interpreter is running really slow, and after comparing the
performance of my interpreter to others, I began thinking of obvious
optimizations.

The First Optimization has actually ALREADY been implemented, that is,
brackets(conditional jumps) work in O(1),that is, there is are no
bracket lookaheads,w hen the jump condiditon is met, it knows EXACTLY
where put the program counter.

In THIS commit, I will begin be skipping some steps, take this code:

+++++++++>>++++++++++++++>>

which is 27 instruction that must be executed! but what we really
should be saying is:

(9+) (2>) (15+) (2>)

Which is 4 instructions executed (because literally every CPU arch can
offset registers by arbitrary values).